+++ composer.json
@@ -1,0 +1,16 @@
+{
+    "name": "user/php_2023",
+    "type": "project",
+    "autoload": {
+        "psr-4": {
+            "User\\Php2023\\": "src/"
+        }
+    },
+    "authors": [
+        {
+            "name": "Evgeniy Evstifeev",
+            "email": "evstifeevea@gmail.com"
+        }
+    ],
+    "require": {}
+}
